+++ src/components/labs/workflow-builder.tsx
@@ -1,3 +1,0 @@
-"use client";
-
-export * from "./workflow-builder/impl";